clean onChange plugin
